+++ app/playlist/loader.h
@@ -3,8 +3,10 @@
 
 #include "track.h"
 
-#include <QStringList>
 #include <QDir>
+#include <QSet>
+#include <QString>
+#include <QStringList>
 #include <QVector>
 
 namespace Playlist {
@@ -28,7 +30,7 @@ namespace Playlist {
     bool is_single_file() const;
     QStringList files_filter() const;
     bool is_cue(const QString &current_path) const;
-    void remove_tracks_added_from_cue(const QStringList &cues, QVector<Track> &tracks) const;
+    void remove_tracks_added_from_cue(const QSet<QString> &cue_audio_paths, QVector<Track> &tracks) const;
   };
 }